deleted unused files, added more comments to main components and used best practice file naming.
